// src/types/matches.ts
// Generated by https://quicktype.io

export interface Match {
  id: number;
  matchDate: string;
  betOffer: number;
  isDeleted: boolean;
  isMatchStarted: boolean;
  isBeted: boolean;
  isWined: boolean;
  result: string;
  isMatchFinished: boolean;
  deletedAt: null;
  teamA: Team;
  teamB: Team;
}

export interface Team {
  id: number;
  name: string;
  teamImage: string;
  deletedAt: null;
  createdAt: string;
}